Finished fix, separation of message validation from plugin registration
