+++ zambeze/orchestration/plugin_modules/globus/globus.py
@@ -1,7 +1,13 @@
 # Local imports
-from .abstract_plugin import Plugin
-from ..identity import validUUID
-from ..network import externalNetworkConnectionDetected
+from ..abstract_plugin import Plugin
+from .globus_common import (
+        localEndpointExists,
+        globusURISeparator,
+        fileURISeparator,
+        getMappedCollections,
+        getGlobusScopes,
+        )
+from ...network import externalNetworkConnectionDetected
 
 # Third party imports
 from globus_sdk import GlobusError
@@ -11,7 +17,6 @@ import globus_sdk
 from copy import deepcopy
 import os
 from os.path import basename
-from os.path import dirname
 from os.path import exists
 from os.path import isdir
 from os.path import isfile
@@ -21,387 +26,7 @@ from typing import Optional
 import json
 import logging
 import pickle
-import re
 import shutil
-
-
-def localEndpointExists(globus_uuid: str, endpoint_list: list[dict]) -> str:
-    for item in endpoint_list:
-        if item["uuid"].lower() == globus_uuid.lower():
-            return True
-    return False
-
-def localEndpointConfigCheck(config: dict):
-    # Check that the UUIDs are correct
-    if "local_endpoints" in config:
-        for local_endpoint in config["local_endpoints"]:
-            if not validUUID(local_endpoint["uuid"]):
-                raise Exception(
-                    f"Invalid uuid detected in plugin: {local_endpoint['uuid']}"
-                )
-            if not exists(local_endpoint["path"]):
-                # Check that the collection path is correct and exists on the local
-                # POSIX filesystem
-                raise Exception(
-                    f"Invalid path detected in plugin: {local_endpoint['path']}"
-                )
-
-        if "default_endpoint" not in config:
-            raise Exception(
-                "'default_endpoint' key value missing from config"
-                " config must have 'default_endpoint' specified if"
-                " local_endpoints are configured."
-            )
-
-        if not validUUID(config["default_endpoint"]):
-            raise Exception(
-                "Invalid uuid detected in plugin for default endpoint: "
-                f" {config['default_endpoint']}"
-            )
-
-        # Make sure that default_endpoint is one of the endpoints that has
-        # been configured
-
-        if not localEndpointExists(
-            config["default_endpoint"], config["local_endpoints"]
-        ):
-            error_msg = f"Invalid default endpoint {config['default_endpoint']}"
-            error_msg = error_msg + " not one of the 'local_endpoints'"
-            error_msg = error_msg + " check your "
-            error_msg = error_msg + " agent.yaml file. Local endpoints are:"
-            error_msg = error_msg + f"\n{config['local_endpoints']}"
-            raise Exception(error_msg)
-
-
-def globusURISeparator(uri: str, default_uuid) -> dict:
-    """Will take a globus URI and break it into its components
-
-    :param uri: the globus uri globus://XXXXX...XXX/path/file.txt
-    :type uri: str
-
-    :Example:
-
-    >>> default_uri = "YYYYZZZZ-YYYY-ZZZZ-YYYY-ZZZZYYYYZZZZ"
-    >>> globus_uri = globus://XXXXYYYY-XXXX-XXXX-XXXX-XXXXYYYYXXXX/path/file.txt
-    >>> uri_components = globusURISeparator(globus_uri, default_uri)
-    >>> print( uri_components.UUID )
-    >>> print( uri_components.path )
-    >>> print( uri_components.file_name )
-    >>> print( uri_components.error_msg )
-
-    The output should be
-
-    >>> XXXXYYYY-XXXX-XXXX-XXXX-XXXXYYYYXXXX
-    >>> /path/
-    >>> file.txt
-
-    :Example: When no endpoint UUID is provided in the URI the default should be used
-
-    >>> default_uri = "YYYYZZZZ-YYYY-ZZZZ-YYYY-ZZZZYYYYZZZZ"
-    >>> globus_uri = globus://path/file.txt
-    >>> uri_components = globusURISeparator(globus_uri, default_uri)
-    >>> print( uri_components.UUID )
-    >>> print( uri_components.path )
-    >>> print( uri_components.file_name )
-    >>> print( uri_components.error_msg )
-
-    The output should be
-
-    >>> YYYYZZZZ-YYYY-ZZZZ-YYYY-ZZZZYYYYZZZZ
-    >>> /path/
-    >>> file.txt
-    """
-    uri = uri.lstrip(" ").rstrip(" ")
-
-    globus_uri_tag = "globus://"
-    # Start by ensuring the start of the uri begins with globus://
-    if not uri.startswith(globus_uri_tag):
-        error_msg = f"Incompatible Globus URI format {uri} must start with "
-        error_msg = error_msg + "globus://"
-        return ("", "", "", error_msg)
-
-    UUID_and_path = uri[len(globus_uri_tag) :]
-    # Replace multiple occurances of // with single /
-    UUID_and_path = re.sub(os.sep + "{2,}", os.sep, UUID_and_path)
-
-    UUID = UUID_and_path[0:36]
-
-    file_and_path = UUID_and_path
-    valid_uuid = default_uuid
-    # Check if the first 36 chars contains os.sep it is probably a file_path
-    # in which case the default uuid should be provided
-    if os.sep not in UUID:
-        if not validUUID(UUID):
-            error_msg = f"Incompatible Globus URI format {uri} must contain 36 "
-            error_msg = error_msg + "character valid UUID of the form "
-            error_msg = error_msg + "XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX"
-            error_msg = error_msg + f" the provided UUID is {UUID} it must also"
-            error_msg = error_msg + " conform to RFC4122"
-            return ("", "", "", error_msg)
-        valid_uuid = UUID
-        file_and_path = UUID_and_path[36:]
-
-    path = dirname(file_and_path)
-
-    if not path.startswith(os.sep):
-        path = os.sep + path
-
-    if not path.endswith(os.sep):
-        path = path + os.sep
-
-    return (valid_uuid, path, basename(file_and_path), "")
-
-
-def fileURISeparator(uri: str) -> dict:
-    """Will take a file URI and break it into its components
-
-    :param uri: File uri should be like file://path/file.txt
-    :type uri: str
-
-    :Example:
-
-    >>> file_uri = file://path/file.txt
-    >>> uri_components = fileURISeparator(file_uri)
-    >>> print( uri_components.path )
-    >>> print( uri_components.file_name )
-    >>> print( uri_components.error_msg )
-
-    The output should be
-
-    >>> /path/
-    >>> file.txt
-    """
-    uri = uri.lstrip(" ").rstrip(" ")
-
-    file_uri_tag = "file://"
-    # Start by ensuring the start of the uri begins with globus://
-    if not uri.startswith(file_uri_tag):
-        error_msg = f"Incompatible file URI format {uri} must start with "
-        error_msg = error_msg + "file://"
-        return ("", "", "", error_msg)
-
-    file_and_path = uri[len(file_uri_tag) :]
-    path = dirname(file_and_path)
-
-    if not path.startswith(os.sep):
-        path = os.sep + path
-
-    if not path.endswith(os.sep):
-        path = path + os.sep
-
-    return (path, basename(file_and_path), "")
-
-
-def checkTransferEndpoint(action_package: dict) -> (bool, str):
-    """Makes sure each item to be transferred has the correct format
-
-    :param action_package: the package that contains instructions for
-    transferring files.
-    :type action_package: dict
-
-    :Example:
-
-    >>> "items": [
-    >>>     {
-    >>>          "source": "globus://XXXXXXXX-...-XXXXXXXXXXXX/file1/txt"
-    >>>          "destination": "globus://YYYYYYYY-...-YYYYYYYYYYYY/dest/file1/txt"
-    >>>     },
-    >>>     {
-    >>>          "source": "globus://XXXXXXXX-...-XXXXXXXXXXXX/file2/txt"
-    >>>          "destination": "globus://YYYYYYYY-...-YYYYYYYYYYYY/dest/file2/txt"
-    >>>     }
-    >>> ]
-
-    """
-    for item in action_package["items"]:
-        if "source" not in item:
-            return False, "'source' missing from 'items' in 'transfer'"
-        else:
-            valid, msg = checkEndpoint(item["source"], ["globus"])
-            if not valid:
-                return (False, "Error in source\n" + msg)
-
-        if "destination" not in item:
-            return False, "'destination' missing from 'items' in 'transfer'"
-        else:
-            valid, msg = checkEndpoint(item["destination"], ["globus"])
-            if not valid:
-                return (False, "Error in destination\n" + msg)
-    return True, ""
-
-
-def getMappedCollections(config: dict) -> list[str]:
-    """Returns a list of the UUIDs that are mapped collections
-
-    :param config: Indicates where on the local filesystem the Globus collection
-    is located.
-    :type config: dict
-
-    :Example:
-
-    >>> config = {
-    >>>     "local_endpoints": [
-    >>>         { "UUID": "XXXX...XXXX", "path": "/here/file", "type": "guest"},
-    >>>         { "UUID": "YYYY...YYYY", "path": "/there/file2", "type": "mapped"}
-    >>>     ]
-    >>> }
-    >>> mapped_coll = getMappedCollections(config)
-    >>> print(mapped_coll)
-    >>> # Single entry would be printed in this case
-    >>> # ["YYYY...YYYY"]
-    """
-    mapped_collections = []
-    if "local_endpoints" in config:
-        for local_endpoint in config["local_endpoints"]:
-            if local_endpoint["type"] == "mapped":
-                mapped_collections.append(local_endpoint["uuid"])
-
-    return mapped_collections
-
-
-def is_36char(item: str):
-    return len(item) == 36
-
-
-def getGlobusScopes(mapped_collections: list[str]) -> str:
-    """Get the globus scopes needed to access the mapped collections
-
-    :param mapped_collections: This should contain the UUIDs of only mapped
-        collections
-    :type mapped_collections: list[str]
-
-    For globus mapped collections, explicit scope has to be requested to
-    access the collection.
-
-    :Example: First example empty mapped collection list
-
-    >>> mapped_coll = [""]
-    >>> scopes = getGlobusScopes(mapped_coll)
-    >>> print(scopes)
-    >>> # Will print general scopes
-    >>> # urn:globus:auth:scope:transfer.api.globus.org:all
-
-    :Example: Second example
-
-    >>> mapped_coll = ["XXXX...XXXX", "YYYY...YYYY"]
-    >>> scopes = getGlobusScopes(mapped_coll)
-    >>> print(scopes)
-    >>> # Will print general scopes
-    >>> # urn:globus:auth:scope:transfer.api.globus.org:all[*https://auth.globus.org/scopes/XXXX...XXXX/data_access *https://auth.globus.org/scopes/YYYY...YYYY/data_access] # noqa: E501
-    """
-    # Clean the UUIDs in the list make sure they are all 36 characters
-    mapped_collections = list(filter(is_36char, mapped_collections))
-
-    scopes = "urn:globus:auth:scope:transfer.api.globus.org:all"
-    if len(mapped_collections):
-        scopes = scopes + "["
-        index = 1
-        for mapped_collection in mapped_collections:
-            scopes = (
-                scopes
-                + f"*https://auth.globus.org/scopes/{mapped_collection}/data_access"
-            )
-            if index < len(mapped_collections):
-                scopes = scopes + " "
-                index = index + 1
-        scopes = scopes + "]"
-    return scopes
-
-
-def checkEndpoint(item: str, supported_types: list[str]) -> (bool, str):
-    """Check that the approprite keys and values exist in the endpoint
-
-    :param item: uri
-    :type item: str
-    :param supported_types: Supported types, defines what values are allowed
-    :type supported_types: list[str]
-    :rtype: bool
-
-    This function will return False if the item provided is an inappropriate value.
-        * type - only two types are currently supported
-        ["globus", "file"]
-
-    :Example:
-
-    >>> item = "globus:://XXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX/file1.txt",
-    >>> valid, msg = checkEndpoint(item)
-    >>> assert valid
-    """
-    # Only "globus relative" path type supported
-    for supported_type in supported_types:
-        if item.startswith(supported_type + "://"):
-            return True, ""
-
-    return (
-        False,
-        f"Missing {item} not in supported types \
-                {supported_types} uri should start with i.e. globus://",
-    )
-
-
-def checkAllItemsHaveValidEndpoints(
-    items: list[dict],
-    supported_source_path_types: list[str],
-    supported_destination_path_types: list[str],
-) -> (bool, str):
-    """Check that all items that are too be moved are schematically correct
-
-    :return: Returns true if the schema of the items is valid and false otherwise
-    :rtype: bool
-
-    :Example: Input
-
-    Provided a list of items to be moved
-
-    >>> "items": [
-    >>>       {
-    >>>           "source": "globus://XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX/file1.txt",
-    >>>           "destination": "globus://YYYYYYYY-...-YYYYYYYYYYYY/dest/file1.txt"
-    >>>       },
-    >>>       {
-    >>>           "source": "globus://XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX/file2.txt",
-    >>>           "destination": "globus://YYYYYYYY-...-YYYYYYYYYYYY/dest/file2.txt"
-    >>>       }
-    >>> ]
-
-    "Example: Input 2
-
-    >>> "items": [
-    >>>     {
-    >>>         "source": "file://file1.txt",
-    >>>         "destination": "globus://YYYYYYYY-YYYY-YYYY-YYYY-YYYYYYYYYYYY/file1.txt"
-    >>>     },
-    >>>     {
-    >>>         "source": "file://file2.txt",
-    >>>         "destination": "globus://YYYYYYYY-YYYY-YYYY-YYYY-YYYYYYYYYYYY/file2.txt"
-    >>>     }
-    >>> ]
-
-
-    >>> supported_source_path_types = ["file"]
-    >>> supported_destination_path_types = ["globus"]
-    >>> assert checkAllItemsHaveValidEndpoints(
-    >>>     items,
-    >>>     supported_source_path_types,
-    >>>     supported_destination_path_types)
-    """
-    for item in items:
-        if "source" not in item:
-            return False, "Item is missing source key"
-        if "destination" not in item:
-            return False, "Item is missing destination key"
-
-        valid, msg = checkEndpoint(item["source"], supported_source_path_types)
-        if not valid:
-            return False, "Invalid source\n" + msg
-
-        valid, msg = checkEndpoint(
-            item["destination"], supported_destination_path_types
-        )
-        if not valid:
-            return False, "Invalid destination\n" + msg
-
-    return True, ""
 
 
 class Globus(Plugin):
@@ -779,43 +404,6 @@ class Globus(Plugin):
 
             shutil.copyfile(source_path, destination_path)
 
-
-    def __runTransferValidationCheck(self, action_package: dict) -> (bool, str):
-        """Checks to ensure that the action_package has the right format and
-        checks for errors.
-
-        :Example:
-
-        >>> {
-        >>>     "type": "synchronous",
-        >>>     "items": [
-        >>>         {
-        >>>             "source": "globus://XXXXXXXX-XX...X-XXXXXXXXXXXX/file1.txt",
-        >>>             "destination": "globus://YYYYYY...-YYYYYYYYYYYY/dest/file1.txt"
-        >>>         },
-        >>>         {
-        >>>             "source": "globus://XXXXXXXX-...XXX-XXXXXXXXXXXX/file2.txt",
-        >>>             "destination": "globus://YYYY...Y-YYYYYYYYYYYY/dest/file2.txt"
-        >>>         }
-        >>>     ]
-        >>> }
-        """
-
-        required_keys = ["type", "items"]
-        for required in required_keys:
-            if required not in action_package:
-                return False, f"{required} key missing from 'transfer' action."
-
-        if action_package["type"] != "synchronous":
-            if action_package["type"] != "asynchronous":
-                return (
-                    False,
-                    "Unsupported 'type' detected. Supported types are \
-            synchronous and asynchronous you have specified {action_package['type']}",
-                )
-
-        return checkTransferEndpoint(action_package)
-
     def __runTransferSanityCheck(self, action_package: dict) -> (bool, str):
         """Checks to ensure that the action_package has the right format and
         checks for errors.
@@ -844,29 +432,6 @@ class Globus(Plugin):
 
         return True, ""
 
-    def __runMoveToGlobusValidationCheck(self, action_package: dict) -> (bool, str):
-        supported_source_path_types = ["file"]
-        supported_destination_path_types = ["globus"]
-
-        valid, msg = checkAllItemsHaveValidEndpoints(
-            action_package["items"],
-            supported_source_path_types,
-            supported_destination_path_types,
-        )
-
-        if valid:
-            for item in action_package["items"]:
-                globus_sep_uri = globusURISeparator(
-                    item["destination"], self.__default_endpoint
-                )
-                if not validUUID(globus_sep_uri[0]):
-                    error_msg = f"Invalid uuid dectected in \
-                                'move_from_globus_collection' item: {item} \nuuid: \
-                                {globus_sep_uri[0]}"
-                    return (False, error_msg)
-
-        return (valid, msg)
-
     def __runMoveToGlobusSanityCheck(self, action_package: dict) -> (bool, str):
         for item in action_package["items"]:
             globus_sep_uri = globusURISeparator(
@@ -885,72 +450,6 @@ class Globus(Plugin):
                 return False, f"Item does not exist {file_path}"
 
         return (True, "")
-
-    def __runGetTaskStatusValidationCheck(self, action_package: dict) -> (bool, str):
-        """Checks that the get_task_status action is correctly configured
-
-        :Example:
-
-        >>> action_package = {
-        >>>     "task_id": "XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX"
-        >>> }
-        """
-        if "task_id" not in action_package:
-            return (False, "Missing 'task_id' in get_task_status action")
-
-        if not validUUID(action_package["task_id"]):
-            return (
-                False,
-                f"Invalid 'task_id' detected in 'get_task_status': \
-            {action_package['task_id']}",
-            )
-        return (True, "")
-
-    def __runMoveFromGlobusValidationCheck(self, action_package: dict) -> (bool, str):
-        """Run a sanity check for the action "move_from_globus_collection"
-
-        return: Will return true if the sanity check passes false otherwise
-
-        Example:
-
-        >>> action_package = {
-        >>>    "source_host_name": "",
-        >>>    "destination_collection_UUID": "",
-        >>>    "items": [
-        >>>           {
-        >>>               "source": "globus://XXXXXXXX-...X-XXXXXXXXXXXX/file1.txt"
-        >>>               "destination": "file://file1.txt",
-        >>>           },
-        >>>           {
-        >>>               "source": "globus://XXXXXXXX-X...XXX-XXXXXXXXXXXX/file2.txt"
-        >>>               "destination": "file://file2.txt",
-        >>>           }
-        >>>    ]
-        >>> }
-        >>> assert self.__runMoveFromGlobusSanityCheck(action_package)
-        """
-
-        supported_source_path_types = ["globus"]
-        supported_destination_path_types = ["file"]
-
-        valid, msg = checkAllItemsHaveValidEndpoints(
-            action_package["items"],
-            supported_source_path_types,
-            supported_destination_path_types,
-        )
-
-        if valid:
-            for item in action_package["items"]:
-                globus_sep_uri = globusURISeparator(
-                    item["source"], self.__default_endpoint
-                )
-                if not validUUID(globus_sep_uri[0]):
-                    error_msg = f"Invalid uuid dectected in \
-                                'move_from_globus_collection' item: {item} \nuuid: \
-                                {globus_sep_uri[0]}"
-                    return (False, error_msg)
-
-        return (valid, msg)
 
     def __runMoveFromGlobusSanityCheck(self, action_package: dict) -> (bool, str):
         """Run a sanity check for the action "move_from_globus_collection"
@@ -1077,54 +576,7 @@ class Globus(Plugin):
 
         self.__configured = True
 
-    def messageTemplate(self, args=None) -> dict:
 
-        if args is None or args == "transfer":
-            return {"transfer": {
-                        "type": "synchronous",
-                        "items": [
-                            {
-                                "source": "globus://XXXXXXXX...X-XXXXXXXX/file1.txt",
-                                "destination": "globus://YYY...YYYYYYYY/dest/file1.txt"
-                            },
-                            {
-                                "source": "globus://XXXXXXXX-...XXXXXXXXXXXX/file2.txt",
-                                "destination": "globus://YYYY...YYYYYYYY/dest/file2.txt"
-                            }
-                        ]
-                        }
-                    }
-        elif args == "move_to_globus_collection":
-            return {"move_to_globus_collection": {
-                "items": [
-                    {
-                        "source": "file://file1.txt",
-                        "destination": "globus://YYYYY...YY-YYYYYYYYYYYY/file1.txt"
-                        },
-                    {
-                        "source": "file://file2.txt",
-                        "destination": "globus://YYYYY...Y-YYYYYYYYYYYY/file2.txt"
-                        }
-                    ]
-                }
-            }
-        elif args == "move_from_globus_collection":
-            return {"move_from_globus_collection": {
-                "items": [
-                    {
-                        "source": "globus://XXXXXXXX-XX...XXXXXXXXXX/file1.txt",
-                        "destination": "file://file1.txt",
-                        },
-                    {
-                        "source": "globus://XXXXXXXX-XX...XXXXXXXXXXX/file2.txt",
-                        "destination": "file://file2.txt",
-                        }
-                    ]
-                }
-            }
-        else:
-            raise Exception("Unrecognized argument provided, cannot generate "
-                            "messageTemplate")
 
     @property
     def supportedActions(self) -> list[str]:
@@ -1176,116 +628,6 @@ class Globus(Plugin):
         information["hostname"] = self.__hostname
         information["configured"] = self.__configured
         return information
-
-    def validateMessage(self, arguments: list[dict]) -> list:
-        """Checks the input argument for errors
-
-        Cycle through the items in the argument and checks if this instance
-        can execute them. This method should be called before process with
-        the same argument.
-
-        Example 1
-
-        >>> arguments = [
-        >>>   { "transfer":
-        >>>       {
-        >>>           "type": "synchronous",
-        >>>           "items": [
-        >>>                 {
-        >>>                     "source": "globus://XXXXXXXX...X-XXXXXXXX/file1.txt",
-        >>>                     "destination": "globus://YYY...YYYYYYYY/dest/file1.txt"
-        >>>                 },
-        >>>                 {
-        >>>                     "source": "globus://XXXXXXXX-...XXXXXXXXXXXX/file2.txt",
-        >>>                     "destination": "globus://YYYY...YYYYYYYY/dest/file2.txt"
-        >>>                 }
-        >>>           ]
-        >>>       }
-        >>>   }
-        >>> ]
-
-        Example 2
-
-        >>> arguments = [
-        >>>   { "move_to_globus_collection": {
-        >>>       "items": [
-        >>>           {
-        >>>               "source": "file://file1.txt",
-        >>>               "destination": "globus://YYYYY...YY-YYYYYYYYYYYY/file1.txt"
-        >>>           },
-        >>>           {
-        >>>               "source": "file://file2.txt",
-        >>>               "destination": "globus://YYYYY...Y-YYYYYYYYYYYY/file2.txt"
-        >>>           }
-        >>>       ]
-        >>>   }
-        >>> ]
-
-        Example 3
-
-        >>> arguments = [
-        >>>   { "move_from_globus_collection": {
-        >>>       "items": [
-        >>>           {
-        >>>               "source": "globus://XXXXXXXX-XX...XXXXXXXXXX/file1.txt"
-        >>>               "destination": "file://file1.txt",
-        >>>           },
-        >>>           {
-        >>>               "source": "globus://XXXXXXXX-XX...XXXXXXXXXXX/file2.txt"
-        >>>               "destination": "file://file2.txt",
-        >>>           }
-        >>>       ]
-        >>>   }
-        >>> ]
-        """
-        checks = []
-        # Here we are cycling a list of dicts
-        for index in range(len(arguments)):
-            for action in arguments[index]:
-                # Check if the action is supported
-                if self.__supported_actions[action] is False:
-                    checks.append({action: (False, "action is not supported.")})
-                    continue
-
-                if action == "transfer":
-                    # Any agent with the globus plugin can submit a job to globus if it
-                    # has access to the globus cloud
-                    checks.append(
-                        {
-                            action: self.__runTransferValidationCheck(
-                                arguments[index][action]
-                            )
-                        }
-                    )
-
-                elif action == "move_to_globus_collection":
-                    checks.append(
-                        {
-                            action: self.__runMoveToGlobusValidationCheck(
-                                arguments[index][action]
-                            )
-                        }
-                    )
-
-                elif action == "move_from_globus_collection":
-                    checks.append(
-                        {
-                            action: self.__runMoveFromGlobusValidationCheck(
-                                arguments[index][action]
-                            )
-                        }
-                    )
-                elif action == "get_task_status":
-                    checks.append(
-                        {
-                            action: self.__runGetTaskStatusValidationCheck(
-                                arguments[index][action]
-                            )
-                        }
-                    )
-                else:
-                    checks.append({action: (False, "Unrecognized action keyword")})
-        return checks
 
     def check(self, arguments: list[dict]) -> list[dict]:
         """Checks the input argument for errors
