+++ zambeze/orchestration/plugins.py
@@ -11,6 +11,7 @@ from __future__ import annotations
 
 # Local imports
 from .plugin_modules.abstract_plugin import Plugin
+from .plugin_modules.abstract_plugin_message_helper import PluginMessageHelper
 
 # Standard imports
 from copy import deepcopy
@@ -59,21 +60,46 @@ class Plugins:
         )
         self.__module_names = []
         self._plugins = {}
+        self._plugin_message_helpers = {}
         self.__registerPlugins()
+        self.__registerPluginHelpers()
 
     def __registerPlugins(self) -> None:
         """Will register all the plugins provided in the plugin_modules folder"""
         plugin_path = [str(Path(__file__).resolve().parent) + "/plugin_modules"]
         for importer, module_name, ispkg in pkgutil.walk_packages(path=plugin_path):
-            if module_name != "abstract_plugin" and module_name != "__init__":
+            if module_name != "abstract_plugin" and \
+               module_name != "__init__" and \
+               module_name != "abstract_plugin_message_helper":
                 self.__module_names.append(module_name)
         self.__logger.debug(f"Registered Plugins: {', '.join(self.__module_names)}")
+
+    def __registerPluginHelpers(self) -> None:
+        for module_name in self.__module_names:
+            # Registering plugin message validators
+            module = import_module(
+                "zambeze.orchestration.plugin_modules."
+                f"{module_name}.{module_name}_message_helper"
+            )
+            for attribute_name in dir(module):
+                potential_plugin_message_helper = getattr(module, attribute_name)
+                if isclass(potential_plugin_message_helper):
+                    if (
+                        issubclass(potential_plugin_message_helper, PluginMessageHelper)
+                        and attribute_name != "PluginMessageHelper"
+                    ):
+                        self._plugin_message_helpers[attribute_name.lower()] = \
+                                potential_plugin_message_helper(
+                            logger=self.__logger
+                        )
+
 
     def messageTemplate(self, plugin_name: str, args) -> dict:
         """Will return a template of the message body that is needed to execute
         an activity using the plugin"""
 
-        message_template = self._plugins[plugin_name].messageTemplate(args)
+        message_template = \
+                self._plugin_message_helpers[plugin_name].messageTemplate(args)
         message_template["plugin"] = plugin_name
         return message_template
 
@@ -138,9 +164,10 @@ class Plugins:
         This will just configure the "shell" plugin
         """
         for module_name in self.__module_names:
+            # Registering plugins
             if module_name in config.keys() and module_name in self.__module_names:
                 module = import_module(
-                    f"zambeze.orchestration.plugin_modules.{module_name}"
+                    f"zambeze.orchestration.plugin_modules.{module_name}.{module_name}"
                 )
                 for attribute_name in dir(module):
                     potential_plugin = getattr(module, attribute_name)
@@ -266,7 +293,7 @@ class Plugins:
         >>> # {
         """
         check_results = {}
-        if plugin_name not in self._plugins.keys():
+        if plugin_name not in self._plugin_message_helpers.keys():
             check_results[plugin_name] = [
                 {"configured": (False, f"{plugin_name} is not configured.")}
             ]
@@ -283,7 +310,7 @@ class Plugins:
                 )
         else:
             check_results[plugin_name] = \
-                self._plugins[plugin_name].validateMessage([arguments])
+                self._plugin_message_helpers[plugin_name].validateMessage([arguments])
         return check_results
 
     def check(self, plugin_name: str, arguments: dict) -> PluginChecks:
